EES-5095 Refactor related functions into `ProcessInitialDataSetVersionFunction`
