+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Tests/Functions/ProcessInitialDataSetVersionFunctionTests.cs
@@ -1,6 +1,11 @@
+using GovUk.Education.ExploreEducationStatistics.Common.Tests.Extensions;
+using GovUk.Education.ExploreEducationStatistics.Public.Data.Model;
+using GovUk.Education.ExploreEducationStatistics.Public.Data.Model.Database;
+using GovUk.Education.ExploreEducationStatistics.Public.Data.Model.Tests.Fixtures;
 using GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Functions;
 using GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Model;
 using Microsoft.DurableTask;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using Moq;
@@ -28,7 +33,7 @@ public class ProcessInitialDataSetVersionFunctionTests(ProcessorFunctionsIntegra
             mockOrchestrationContext
                 .Setup(context =>
                     context.CallActivityAsync(
-                        nameof(CompleteProcessingFunction.CompleteProcessing),
+                        nameof(ProcessInitialDataSetVersionFunction.CompleteProcessing),
                         processingContext,
                         null))
                 .Returns(Task.CompletedTask);
@@ -54,7 +59,7 @@ public class ProcessInitialDataSetVersionFunctionTests(ProcessorFunctionsIntegra
             mockOrchestrationContext
                 .Setup(context =>
                     context.CallActivityAsync(
-                        nameof(CompleteProcessingFunction.CompleteProcessing),
+                        nameof(ProcessInitialDataSetVersionFunction.CompleteProcessing),
                         processingContext,
                         null))
                 .Throws<Exception>();
@@ -62,7 +67,7 @@ public class ProcessInitialDataSetVersionFunctionTests(ProcessorFunctionsIntegra
             mockOrchestrationContext
                 .Setup(context =>
                     context.CallActivityAsync(
-                        nameof(HandleProcessingFailureFunction.HandleProcessingFailure),
+                        nameof(ProcessInitialDataSetVersionFunction.HandleProcessingFailure),
                         processingContext,
                         null))
                 .Returns(Task.CompletedTask);
@@ -95,6 +100,115 @@ public class ProcessInitialDataSetVersionFunctionTests(ProcessorFunctionsIntegra
                 .Returns(isReplaying);
 
             return mock;
+        }
+    }
+
+    public class HandleProcessingFailureTests(ProcessorFunctionsIntegrationTestFixture fixture)
+        : ProcessInitialDataSetVersionFunctionTests(fixture)
+    {
+        [Fact]
+        public async Task Success()
+        {
+            var instanceId = Guid.NewGuid();
+
+            DataSet dataSet = DataFixture.DefaultDataSet();
+
+            await AddTestData<PublicDataDbContext>(context => context.DataSets.Add(dataSet));
+
+            DataSetVersion dataSetVersion = DataFixture
+                .DefaultDataSetVersion()
+                .WithDataSet(dataSet)
+                .WithStatus(DataSetVersionStatus.Processing)
+                .WithImports(
+                    () => DataFixture
+                        .DefaultDataSetVersionImport()
+                        .WithInstanceId(instanceId)
+                        .WithStage(DataSetVersionImportStage.Pending)
+                        .Generate(1)
+                )
+                .FinishWith(dsv => dsv.DataSet.LatestDraftVersion = dsv);
+
+            await AddTestData<PublicDataDbContext>(
+                context =>
+                {
+                    context.DataSetVersions.Add(dataSetVersion);
+                    context.DataSets.Update(dataSet);
+                }
+            );
+
+            var function = GetRequiredService<ProcessInitialDataSetVersionFunction>();
+            await function.HandleProcessingFailure(
+                new ProcessInitialDataSetVersionContext
+                {
+                    DataSetVersionId = dataSetVersion.Id
+                },
+                instanceId,
+                CancellationToken.None
+            );
+
+            await using var publicDataDbContext = GetDbContext<PublicDataDbContext>();
+
+            var savedDataSetVersion = await publicDataDbContext.DataSetVersions
+                .Include(dsv => dsv.Imports)
+                .FirstAsync(dsv => dsv.Id == dataSetVersion.Id);
+
+            Assert.Equal(DataSetVersionStatus.Failed, savedDataSetVersion.Status);
+
+            var savedImport = savedDataSetVersion.Imports.Single();
+            Assert.Equal(DataSetVersionImportStage.Pending, savedImport.Stage);
+            savedImport.Completed.AssertUtcNow();
+        }
+    }
+
+    public class CompleteProcessingTests(ProcessorFunctionsIntegrationTestFixture fixture)
+        : ProcessorFunctionsIntegrationTest(fixture)
+    {
+        [Fact]
+        public async Task Success()
+        {
+            var instanceId = Guid.NewGuid();
+
+            DataSet dataSet = DataFixture
+                .DefaultDataSet();
+
+            await AddTestData<PublicDataDbContext>(context => context.DataSets.Add(dataSet));
+
+            DataSetVersion dataSetVersion = DataFixture
+                .DefaultDataSetVersion()
+                .WithDataSet(dataSet)
+                .WithStatus(DataSetVersionStatus.Processing)
+                .WithImports(() => DataFixture
+                    .DefaultDataSetVersionImport()
+                    .WithInstanceId(instanceId)
+                    .WithStage(DataSetVersionImportStage.Pending)
+                    .Generate(1))
+                .FinishWith(dsv => dsv.DataSet.LatestDraftVersion = dsv);
+
+            await AddTestData<PublicDataDbContext>(context =>
+            {
+                context.DataSetVersions.Add(dataSetVersion);
+                context.DataSets.Update(dataSet);
+            });
+
+            var function = GetRequiredService<ProcessInitialDataSetVersionFunction>();
+            await function.CompleteProcessing(new ProcessInitialDataSetVersionContext
+                {
+                    DataSetVersionId = dataSetVersion.Id
+                },
+                instanceId,
+                CancellationToken.None);
+
+            await using var publicDataDbContext = GetDbContext<PublicDataDbContext>();
+
+            var savedDataSetVersion = await publicDataDbContext.DataSetVersions
+                .Include(dsv => dsv.Imports)
+                .FirstAsync(dsv => dsv.Id == dataSetVersion.Id);
+
+            Assert.Equal(DataSetVersionStatus.Draft, savedDataSetVersion.Status);
+
+            var savedImport = savedDataSetVersion.Imports.Single();
+            Assert.Equal(DataSetVersionImportStage.Completing, savedImport.Stage);
+            savedImport.Completed.AssertUtcNow();
         }
     }
 }
