+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Tests/ProcessorFunctionsIntegrationTest.cs
@@ -61,9 +61,7 @@ public class ProcessorFunctionsIntegrationTestFixture : FunctionsIntegrationTest
     {
         return
         [
-            typeof(CompleteProcessingFunction),
             typeof(CreateInitialDataSetVersionFunction),
-            typeof(HandleProcessingFailureFunction),
             typeof(ProcessInitialDataSetVersionFunction),
         ];
     }
