+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Processor/Functions/ProcessInitialDataSetVersionFunction.cs
@@ -1,11 +1,14 @@
+using GovUk.Education.ExploreEducationStatistics.Public.Data.Model;
+using GovUk.Education.ExploreEducationStatistics.Public.Data.Model.Database;
 using GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Model;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.DurableTask;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 
 namespace GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Functions;
 
-public class ProcessInitialDataSetVersionFunction
+public class ProcessInitialDataSetVersionFunction(PublicDataDbContext publicDataDbContext)
 {
     [Function(nameof(ProcessInitialDataSetVersion))]
     public async Task ProcessInitialDataSetVersion(
@@ -28,11 +31,11 @@ public class ProcessInitialDataSetVersionFunction
             // Import data to DuckDb
             // Export to Parquet files
 
-            await context.CallActivityAsync(nameof(CompleteProcessingFunction.CompleteProcessing), input);
+            await context.CallActivityAsync(nameof(CompleteProcessing), input);
 
             logger.LogInformation(
                 "Activity '{ActivityName}' completed (InstanceId={InstanceId}, DataSetVersionId={DataSetVersionId})",
-                nameof(CompleteProcessingFunction.CompleteProcessing),
+                nameof(CompleteProcessing),
                 context.InstanceId,
                 input.DataSetVersionId);
         }
@@ -43,7 +46,59 @@ public class ProcessInitialDataSetVersionFunction
                 context.InstanceId,
                 input.DataSetVersionId);
 
-            await context.CallActivityAsync(nameof(HandleProcessingFailureFunction.HandleProcessingFailure), input);
+            await context.CallActivityAsync(nameof(HandleProcessingFailure), input);
         }
+    }
+
+    [Function(nameof(CompleteProcessing))]
+    public async Task CompleteProcessing(
+        [ActivityTrigger] ProcessInitialDataSetVersionContext input,
+        Guid instanceId,
+        CancellationToken cancellationToken)
+    {
+        var dataSetVersion = await GetDataSetVersion(
+            dataSetVersionId: input.DataSetVersionId,
+            instanceId: instanceId,
+            cancellationToken);
+
+        var dataSetVersionImport = dataSetVersion.Imports.Single(i => i.InstanceId == instanceId);
+
+        dataSetVersionImport.Stage = DataSetVersionImportStage.Completing;
+        await publicDataDbContext.SaveChangesAsync(cancellationToken);
+
+        // Any additional logic to tidy up after importing will be added here
+
+        dataSetVersion.Status = DataSetVersionStatus.Draft;
+        dataSetVersionImport.Completed = DateTimeOffset.UtcNow;
+        await publicDataDbContext.SaveChangesAsync(cancellationToken);
+    }
+
+    [Function(nameof(HandleProcessingFailure))]
+    public async Task HandleProcessingFailure(
+        [ActivityTrigger] ProcessInitialDataSetVersionContext input,
+        Guid instanceId,
+        CancellationToken cancellationToken)
+    {
+        var dataSetVersion = await GetDataSetVersion(
+            dataSetVersionId: input.DataSetVersionId,
+            instanceId: instanceId,
+            cancellationToken);
+
+        dataSetVersion.Status = DataSetVersionStatus.Failed;
+
+        var dataSetVersionImport = dataSetVersion.Imports.Single(i => i.InstanceId == instanceId);
+        dataSetVersionImport.Completed = DateTimeOffset.UtcNow;
+
+        await publicDataDbContext.SaveChangesAsync(cancellationToken);
+    }
+
+    private async Task<DataSetVersion> GetDataSetVersion(
+        Guid dataSetVersionId,
+        Guid instanceId,
+        CancellationToken cancellationToken)
+    {
+        return await publicDataDbContext.DataSetVersions
+            .Include(dsv => dsv.Imports.Where(i => i.InstanceId == instanceId))
+            .SingleAsync(dsv => dsv.Id == dataSetVersionId, cancellationToken: cancellationToken);
     }
 }
